# backend/Swagger.yaml
openapi: 3.0.1
info:
  title: Delilah Resto
  description: API Backend de un restaurante de comida típica
  contact:
    email: dmesa@gmail.com
  version: 1.0.0
...
servers:
  - url: https://localhost:3000/
  - url: http://localhost:3000/
...
paths:
...
  /addOrder/:id:
    post:
      tags:
        - Orders
      summary: Agrega una orden
      description: Agrega órdenes de acuerdo a los productos seleccionados
      operationId: addOrder
      requestBody:
        description: Order to create
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/addOrder"
        required: true
      responses:
        200:
          description: Orden agregada con éxito
        400:
          description: Información incompleta o petición mal formulada
        default:
          $ref: "#/components/responses/Default"
  /updateOrder/:id:
    put:
      tags:
        - Orders
      summary: Actualiza la orden (solo Administrador)
      description: Actualiza toda la información de la orden de acuerdo a la id de la orden
      operationId: updateOrder
      parameters:
        - name: id
          in: query
          example: id=1
          description: Se necesita la identificación de la orden.
          required: true
          schema:
            type: integer
      requestBody:
        description: updateOrder
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  example: Entregado
        required: true
      responses:
        200:
          description: "Orden ${id} fue modificada con éxito"
        400:
          description: No se envió identificación
        401:
          $ref: "#/components/responses/UnauthorizedAdmin"
        403:
          description: El parámetro id (identificación) debe ser valor numérico
        404:
          description: No se encontraron resultados
        default:
          $ref: "#/components/responses/Default"
  /deleteOrder/:id:
    delete:
      tags:
        - Orders
      summary: Eliminar orden (solo Administrador)
      description: Elimina una orden por la identificación.
      parameters:
        - name: id
          in: query
          example: id=1
          description: Necesitamos la identificación de la orden.
          required: true
          schema:
            type: string
      operationId: deleteOrder
      responses:
        200:
          description: "Se eliminó la orden ${id} con éxito"
        400:
          description: No se envio identificación, debe ser un valor numérico
        401:
          $ref: "#/components/responses/UnauthorizedAdmin"
        default:
          $ref: "#/components/responses/Default"
  
#PRODUCTOS
  /getProducts:
    get:
      tags:
        - Products
      summary: Obtiene todos los productos
      description: Devuelve una lista de productos
      operationId: getProducts
      parameters:
        - name: search
          in: query
          description: Busca todos los productos
          schema:
            type: string
      responses:
        200:
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Products'
        400:
          description: Error al buscar la información
        401:
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Default"
...
  /addProducts/:id:      
    post:
      tags:
        - Products
      summary: Agrega un producto (solo Administrador)
      description: Agrega producto a la base de datos
      operationId: addProducts
      requestBody:
        description: Agrega un producto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        201:
          description: Producto agregado correctamente
        400:
          description: Información incompleta
        401:
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Default"
  /updateProducts/:id:  
    put:
      tags:
        - Products
      summary: Actualiza producto (solo Administrador)
      description: Actualiza el estado de un producto por su identificación
      parameters:
        - name: id
          in: query
          example: id=3
          description: Se necesita la identificación del producto.
          required: true
          schema:
            type: string
      operationId: updateProducts
      requestBody:
        description: Modification data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        200:
          description: Actualización satisfatoria
        400:
          description: Información incompleta o petición mal formulada
        401:
          $ref: "#/components/responses/UnauthorizedAdmin"
        default:
          $ref: "#/components/responses/Default"
  /deleteProducts/:id:  
    delete:
      tags:
        - Products
      summary: Elimina un producto (solo Administrador)
      description: Elimina el producto por su identificación
      parameters:
        - name: id
          in: query
          example: id=3
          description: Necesitamos la identificacion del producto.
          required: true
          schema:
            type: string
      operationId: deleteProducts
      responses:
        200:
          description: "Producto eliminado por id ${id}"
        400:
          description: No se puede eliminar producto con id:${id}
        401:
          $ref: "#/components/responses/UnauthorizedAdmin"
        406:
          description: Falta identificación
        default:
          $ref: "#/components/responses/Default"

          
#COMPONENTES
components:
  schemas:
...
    Product:
      type: object
      properties:
        product_id:
          type: integer
          example: 3
        name:
          type: string
          example: Empanadas Antioqueñas
        price:
          type: number
          example: 360
        img_url:
          type: string
          example: https://empanada.jpg
        description:
          type: string
          example: Empanadas de papa acompañadas de ají
        is_disabled:
          type: boolean
          example: 0
    Product2:
      type: object
      properties:
        product_id:
          type: integer
          example: 2
        name:
          type: string
          example: Chicharrón
        price:
          type: number
          example: 300
        img_url:
          type: string
          example: https://via.placeholder.com/256
        description:
          type: string
          example: Entero o picado, servido con patacón y arepa
        is_disabled:
          type: boolean
          example: 0
    Products:
      type: object
      properties:
        1:
          type: object
          $ref: "#/components/schemas/Product"
        2:
          type: object
          $ref: "#/components/schemas/Product2"
    
#schema-Orders_Products
...
    addOrder:
      type: object
      properties:
        payment_method:
          type: string
          example: cash
        products:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: integer
                example: 2
              amount:
                type: integer
                example: 5
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "An error has ocurred: <error>"
      required:
        - error
    Send:
      type: string
      example: "Error message"
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    
      

  # Reused Responses
  responses:
...
    Default:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/Send"
          examples:
            Token Invalidation:
              summary: Unauthorized - Invalid Token
              value: Unauthorized - Invalid Token
            Disabled Account:
              summary: Unauthorized - Account is Disabled
              value: Unauthorized - User account is disabled
    UnauthorizedAdmin:
      description: Unauthorized
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/Send"
          examples:
            Token Invalidation:
              summary: Unauthorized - Invalid Token
              value: Unauthorized - Invalid Token
            Disabled Account:
              summary: Unauthorized - Account is Disabled
              value: Unauthorized - User account is disabled
            Not Admin:
              summary: Unauthorized - Not an admin
              value: Unauthorized - Not an admin
...
security:
  - bearerAuth: []          